Move echo functionality from loader to echoService
